Searches for a value in a sorted array of integers using the Exponential search algorithm

+++ 0x1E-search_algorithms/103-exponential.c
@@ -45,7 +45,7 @@ int binary_srch(int *array, size_t low, size_t high, int value)
 		for (i = low; i <= high; i++)
 			printf("%i%s", array[i], i == high ? "\n" : ", ");
 		if (array[middle] == value)
-			return (middle);
+			return ((int)middle);
 		else if (array[middle] < value)
 			low = middle + 1;
 		else
